1. cast's to avoid warnings on solaris (sparc5) 2. void casts for returns strtol and chown 3. avoid int64, adapt zparser_ttl2int
git-svn-id: file:///svn/nsd/trunk@2878 a26ef69c-88ff-0310-839f-98b793d9c207

+++ dns.c
@@ -490,7 +490,7 @@ rrtype_from_string(const char *name)
 	if (strncasecmp(name, "TYPE", 4) != 0)
 		return 0;
 
-	if (!isdigit(name[4]))
+	if (!isdigit((int)name[4]))
 		return 0;
 
 	/* The rest from the string must be a number.  */
@@ -531,20 +531,19 @@ rrclass_from_string(const char *name)
 
 	if (strlen(name) < 6)
 		return 0;
-	
+
 	if (strncasecmp(name, "CLASS", 5) != 0)
 		return 0;
 
-	if (!isdigit(name[5]))
+	if (!isdigit((int)name[5]))
 		return 0;
-	
+
 	/* The rest from the string must be a number.  */
 	rrclass = strtol(name + 5, &end, 10);
 	if (*end != '\0')
 		return 0;
 	if (rrclass < 0 || rrclass > 65535L)
 		return 0;
-	
-        return (uint16_t) rrclass;
-	
+
+	return (uint16_t) rrclass;
 }
